Overhauled Analysis Manager.py
Implemented functionality to select plots via command line arguments. This is done by passing -p followed by a list of plots separated by commas. Running different plots no longer requires uncommenting and recommenting specific lines of code.

Added wrapper methods for all methods in analysis manager to facilitate calling the methods via passed args.

Removed plot_trackstry.py and plot_tracks_reed_rewrite method completely because they did not work and even if they did work would add no additional functionality.

Combined all functionality of plot_tracks.py and plot tracks method within the analysis manager. Tracks are now plotted differently.  The tracks used to be plotted by pre-processing tracks in the plot_tracks method and writing the results to a file. After the simulation was finished running, the plot_tracks.py file would be run which would load the tracks file and plot the tracks. Now, the analysis manager calls the pre-process_tracks method, which pre-processes the tracks and stores them as instance variables. Then, the plot_tracks method is called, which plots the tracks. The plot_tracks method takes 3 arguments, the tracks that are to be plotted, the title of the plot, and a boolean to determine whether or not to plot the geometry components in addition to the tracks. This eliminates the need to write to files and to run the code in multiple steps and also allows for multiple kinds of tracks to be plotted in one run, improving speed, decreasing memory usage, and adding more funcitonality.